Make highlighting more flexible.
The wizard no longer has a 'highlight color' menu item.  Instead, it
makes two selection: 'wt_hilite' and 'mut_hilite'.  You can change the
color of these selections to change the highlight color for either
residue.  This approach is more flexible but less intuitive.  I'm not
sure which is better.