Don't just align the residue is the focus sele.
A few days ago I added the ability for the user to select a subset of
residues to focus on, in case the two proteins being compared had more
differences than the user was interested in seeing.  I did this by
basically applying the given selection to both the wildtype and mutant
objects.

However, this approach dramatically reduced the number of residues
participating in the sequence alignment, which led to poor mutation
identification.  I tried to address that issue in this commit by always
aligning all the residues in the wildtype and mutant object, and just
filtering the mutations to display by the focus selection.  So far, this
approach seems to be much more robust.