Adding support for non-canonical amino acids
Before, the script treated unrecognized residues as "" (null strings) in the sequence contstructed for alignment.
This resulted in numbering mismatches for mutations after the non-canonical residue, and the wrong residue being
centered and highlighted for disply.

Now, the script treats all non-canonical sequences as "X" when aligning. Mismatches between non-canonicals are given a generic,
BLOSUM-style score of -4 (e.g. "A"->"X"), and matches a score of 5.